tof32

+++ src/activation_layer.rs
@@ -3,13 +3,13 @@ use ndarray::Array2;
 use crate::layer::{Layer, LayerType, SerializedLayer};
 
 pub struct ActivationLayer {
-    input: Array2<f64>,
-    activation: fn(&Array2<f64>) -> Array2<f64>,
-    activation_prime: fn(&Array2<f64>) -> Array2<f64>
+    input: Array2<f32>,
+    activation: fn(&Array2<f32>) -> Array2<f32>,
+    activation_prime: fn(&Array2<f32>) -> Array2<f32>
 }
 
 impl ActivationLayer{
-    pub fn new(activation: fn(&Array2<f64>) -> Array2<f64>, activation_prime:fn(&Array2<f64>) -> Array2<f64>) -> ActivationLayer{
+    pub fn new(activation: fn(&Array2<f32>) -> Array2<f32>, activation_prime:fn(&Array2<f32>) -> Array2<f32>) -> ActivationLayer{
         ActivationLayer{
             input: Array2::default((1, 1)),
             activation,
@@ -19,16 +19,16 @@ impl ActivationLayer{
 }
 
 impl Layer for ActivationLayer {
-    fn forward_propagation(&mut self, input: Array2<f64>) -> Result<Array2<f64>, String> {
+    fn forward_propagation(&mut self, input: Array2<f32>) -> Result<Array2<f32>, String> {
         self.input = input;
         Ok((self.activation)(&self.input))
     }
 
     fn backward_propagation(
         &mut self,
-        output_error: Array2<f64>,
-        _learning_rate: f64,
-    ) -> Result<Array2<f64>, String> {
+        output_error: Array2<f32>,
+        _learning_rate: f32,
+    ) -> Result<Array2<f32>, String> {
         Ok((self.activation_prime)(&self.input) * output_error)
     }
     fn get_type(&mut self) -> crate::layer::LayerType {
