+++ src/fully_connected_layer.rs
@@ -5,10 +5,10 @@ use rand::Rng;
 use crate::layer::{Layer, LayerType, SerializedLayer};
 
 pub struct FullyConnectedLayer {
-    pub weights: Array2<f64>,
-    pub bias: Array2<f64>,
-    input: Array2<f64>,
-    //output: Array2<f64>,
+    pub weights: Array2<f32>,
+    pub bias: Array2<f32>,
+    input: Array2<f32>,
+    //output: Array2<f32>,
 }
 
 impl FullyConnectedLayer {
@@ -18,10 +18,10 @@ impl FullyConnectedLayer {
         // Define the shape of the array
         let shape = (input_size, output_size);
         // Create a new array with the specified shape and fill it with random values
-        let weights: Array2<f64> = Array2::from_shape_fn(shape, |_| (rng.gen::<f64>() - 0.5));
+        let weights: Array2<f32> = Array2::from_shape_fn(shape, |_| (rng.gen::<f32>() - 0.5));
         // Define the shape of the array
         let shape = (1, output_size);
-        let bias: Array2<f64> = Array2::from_shape_fn(shape, |_| rng.gen::<f64>() - 0.5);
+        let bias: Array2<f32> = Array2::from_shape_fn(shape, |_| rng.gen::<f32>() - 0.5);
 
         FullyConnectedLayer {
             weights,
@@ -34,16 +34,16 @@ impl FullyConnectedLayer {
 
 
 impl Layer for FullyConnectedLayer {
-    fn forward_propagation(&mut self, input: Array2<f64>) -> Result<Array2<f64>, String> {
+    fn forward_propagation(&mut self, input: Array2<f32>) -> Result<Array2<f32>, String> {
         self.input = input;
         let mid = self.input.dot(&self.weights);
         Ok(mid.add(&self.bias))
     }
     fn backward_propagation(
         &mut self,
-        output_error: Array2<f64>,
-        learning_rate: f64,
-    ) -> Result<Array2<f64>, String> {
+        output_error: Array2<f32>,
+        learning_rate: f32,
+    ) -> Result<Array2<f32>, String> {
         let input_error = output_error.dot(&self.weights.t());
         let weights_error = self.input.t().dot(&output_error);
         self.weights = &self.weights - (learning_rate * weights_error);
